api/pyproject.toml
[tool.poetry]
name = "codando-api"
version = "0.1.0"
description = ""
authors = ["Bernardo Meirelles <bernardomrl@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
fastapi = "^0.105.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
